requirements.txt
discord.py==2.3.2
lavalink==5.4.0
PyNaCl==1.5.0
